backend/requirements.txt
# ============================================================
# Core Requirements
# ============================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ============================================================
# GraphQL
# ============================================================
strawberry-graphql[asgi]==0.214.0
graphql-core==3.2.3

# ============================================================
# Database
# ============================================================
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# ============================================================
# ML & AI
# ============================================================
# Traditional ML
torch==2.1.0
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.26.2
pandas==2.1.3

# AI Reasoning (NEW)
langchain==0.1.0
ollama-python==0.1.0

# Vector Embeddings (NEW)
sentence-transformers==2.2.2
qdrant-client==1.7.0

# ============================================================
# Real-Time & Streaming (NEW)
# ============================================================
pathway==0.7.0
kafka-python==2.0.2

# ============================================================
# Orbital Mechanics (NASA-grade precision)
# ============================================================
skyfield==1.46  # SGP4/SDP4 propagation, high-precision ephemerides

# ============================================================
# Reinforcement Learning
# ============================================================
gym==0.26.2  # OpenAI Gym for RL environment
stable-baselines3==2.2.1  # PPO agent for maneuver optimization
tensorboard==2.15.1  # Training visualization

# ============================================================
# Caching & Session
# ============================================================
redis==5.0.1
aioredis==2.0.1

# ============================================================
# Authentication & Security
# ============================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Security Testing (NEW)
aiohttp==3.9.1
python-nmap==0.7.1

# ============================================================
# Monitoring & Observability
# ============================================================
prometheus-client==0.19.0
psutil==5.9.6

# ============================================================
# HTTP & Requests
# ============================================================
httpx==0.25.1
requests==2.31.0

# ============================================================
# Space Data Processing
# ============================================================
sgp4==2.23
skyfield==1.46
astropy==5.3.4

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.0
pyyaml==6.0.1
ujson==5.8.0
orjson==3.9.10

# ============================================================
# CORS & Middleware
# ============================================================
python-cors==1.0.0

# ============================================================
# Testing (Development)
# ============================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.0.3
httpx==0.25.1

# ============================================================
# Code Quality (Development)
# ============================================================
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
